Add copying information
I tried to make the copying information very precise and concise. I
specifically tried to avoid saying “dedicated to the public domain”
since (depending on the jurisdiction) the work may be public domain or
may be available under the fallback license.

I specifically chose to use the verb “apply” since CC0 itself uses that
verb to describe the relationship between CC0 and the works that CC0
affects (the Statement of Purpose says “[…]voluntarily elects to apply
CC0 to the Work[…]”).

I had considered saying that CC0 applies to the entire repo, but I
realized that that would make it harder for people to fork the project
and use a different legal tool (they would have to remove or edit the
old commits to not make statements about the rest of the repo).